+++ server/src/main/java/com/linbit/linstor/netcom/PeerTask.java
@@ -318,7 +318,7 @@ public class PeerTask implements Peer
     }
 
     @Override
-    public void fullSyncFailed()
+    public void fullSyncFailed(ApiConsts.ConnectionStatus ignored)
     {
     }